refactor: rename EditorKeyboardShortcutsActions to EditorKeyboardShortcutsActionsManager, move raw_editor_actions to keyboard_shortcuts/editor_keyboard_shortcut_actions_manager.dart
